+++ src/onForumThreadCreate.js
@@ -1,0 +1,26 @@
+// @ts-check
+import { ChannelType } from 'discord.js'
+/**
+ * @typedef {import('./logger.js').Logger} Logger
+ * @typedef {import('discord.js').AnyThreadChannel} AnyThreadChannel
+ */
+
+/**
+ * @param {Logger} logger
+ * @param {AnyThreadChannel} thread
+ * @param {(ownerId: string) => string} messageContent
+ */
+export async function onForumThreadCreate(logger, thread, messageContent) {
+  if (thread.parent?.type !== ChannelType.GuildForum) return
+  if (!thread.ownerId) return
+
+  await thread
+    .send({
+      content: messageContent(thread.ownerId),
+    })
+    .then(it => it.suppressEmbeds())
+
+  logger.info(
+    `"${thread.parent.name}" (${thread.parentId}) で"${thread.name}" (${thread.id}) が作成されました。`
+  )
+}
